make functional and change file structure
