--- datascraper/__init__.py
@@ -1,1 +1,0 @@
-from .datascrapper import DataScrapper
