+++ package/__init__.py
@@ -1,0 +1,3 @@
+__all__ = ['datascraper']
+
+# makes datascraper.py functions available at the package level
