+++ setup.py
@@ -18,7 +18,7 @@ setup(name='mymodule',
       author='Tyler Miguel',
       author_email='miguel_tyler@yahoo.com',
       license='MIT',
-      packages=['datascrapper'],
+      packages=['package'],
       install_requires=['contextlib', 'requests', 'bs4', 'pandas', 'collections', 'os', 'sys'],
       include_package_data=True,
       zip_safe=False)
